+++ lib/src/big_decimal_to_string_converter.dart
@@ -1,0 +1,5 @@
+abstract class BigDecimalToStringConverter {
+  String asWords(double value);
+
+  String asWordsWithCurrency(double value, String currencySymbol);
+}