+++ lib/src/languages/english/english_values.dart
@@ -1,6 +1,7 @@
 import '../../support/base_numbers_builder.dart';
 
 import '../../base_values.dart';
+import '../../currency.dart';
 import '../gender_forms.dart';
 import '../plural_forms.dart';
 import 'english_plural_forms.dart';
@@ -65,6 +66,15 @@ class EnglishValues implements BaseValues {
   @override
   String currency() {
     return "£";
+  }
+
+  @override
+  Currency getCurrency() {
+    return Currency(
+      symbol: "\$",
+      names: ["US dollar", "US dollars"],
+      subunitNames: ["cent", "cents"],
+    );
   }
 
   @override