+++ lib/tradukisto.dart
@@ -3,4 +3,5 @@
 /// More dartdocs go here.
 library;
 
+export 'money_converters.dart';
 export 'value_converters.dart';